+++ src/test/java/chess/service/UserServiceTest.java
@@ -3,18 +3,12 @@ package chess.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import chess.dao.UserDaoInterface;
 import chess.dto.user.UserRequestDto;
-import chess.entity.User;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
+import chess.service.testdao.TestUserDao;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class UserServiceTest {
+public class UserServiceTest {
 
     private UserService userService;
 
@@ -33,29 +27,5 @@ class UserServiceTest {
 
         assertThat(userService.find(name).getName())
             .isEqualTo(name);
-    }
-}
-
-class TestUserDao implements UserDaoInterface {
-
-    private final List<User> users = new ArrayList<>();
-
-    @Override
-    public void insert(final String name) {
-        users.add(
-            new User(users.size(), name, 0, 0, LocalDateTime.now())
-        );
-    }
-
-    @Override
-    public Optional<User> selectByName(final String name) {
-        return users.stream()
-            .filter(user -> Objects.equals(user.getName(), name))
-            .findAny();
-    }
-
-    @Override
-    public void deleteByName(final String name) {
-
     }
 }
